[improving_sync] Adding support to autoreload api if code changes.
