--- src/agent/app.js
@@ -1,8 +1,0 @@
-var express = require('express');
-var app = express();
-
-app.get('/', (req, res) => {
-  res.send('hello world');
-});
-
-export { app };
